apply ODrive native protocol updates to fibre and switch to fibre
