+++ Firmware/MotorControl/odrive_main.h
@@ -88,7 +88,7 @@ inline ENUMTYPE operator ~ (ENUMTYPE a) { return static_cast<ENUMTYPE>(~static_c
 
 
 // ODrive specific includes
-#include <communication/protocol.hpp>
+#include <fibre/protocol.hpp>
 #include <utils.h>
 #include <low_level.h>
 #include <encoder.hpp>
