+++ Firmware/communication/ascii_protocol.hpp
@@ -1,13 +1,9 @@
 #ifndef __ASCII_PROTOCOL_H
 #define __ASCII_PROTOCOL_H
 
-#ifdef __cplusplus
-extern "C" {
-#endif
 
 /* Includes ------------------------------------------------------------------*/
-
-#include "protocol.hpp"
+#include <fibre/protocol.hpp>
 
 #include <stdlib.h>
 #include <stdint.h>
@@ -22,8 +18,5 @@ extern "C" {
 /* Exported functions --------------------------------------------------------*/
 void ASCII_protocol_parse_stream(const uint8_t* buffer, size_t len, StreamSink& response_channel);
 
-#ifdef __cplusplus
-}
-#endif
 
 #endif /* __ASCII_PROTOCOL_H */
