+++ Firmware/fibre/README.md
@@ -74,7 +74,7 @@ Say you want to publish `test_object` so that a remote Fibre node can use it.
             FIBRE_EXPORTS(TestClass,
                   make_protocol_property("property1", &property1),
                   make_protocol_property("property2", &property2),
-                  make_protocol_function("set_both", obj, &TestClass::set_both, "arg1", "arg2")
+                  make_protocol_function("set_both", *obj, &TestClass::set_both, "arg1", "arg2")
             );
       };
       ```
